💥 Fix max packet size calculations
Migration guide:

This change brings the minimum supported version of the Rust compiler up to 1.57.0.

Instead of using the `MAX_ESCAPE_HEADER_BYTES` constant to manually calculate the required output buffer size, this is now exposed as a function on the `FormatDescription` struct. Simply call `max_packet_size()` on your output format description struct.

This functions is a const fn, so it can be used in const contexts. In addition, the `FormatDescription::alac` constructor was also made const, so that the minimum output buffer size can be calculated at compile time.